Refactored code and Reservation/ Offer structure (BMA-7)
