+++ app/src/main/java/com/davidmarian_buzatu/bookster/services/OfferActions.java
@@ -45,144 +45,12 @@ public class OfferActions {
         deleteOffer(offer).addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
             public void onComplete(@NonNull Task<Void> task) {
-                deleteAllReservationsForOffer(offer, context);
+                ReservationActions.getInstance().deleteAllReservationsForOffer(offer, context, mDialog);
             }
         });
 
     }
 
-    private void deleteAllReservationsForOffer(Offer offer, Context context) {
-        List<Reservation> newReservationList = new ArrayList<>();
-        getAllReservations().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                if (task.isSuccessful() && task.getResult() != null) {
-                    for (QueryDocumentSnapshot doc : task.getResult()) {
-                        // we get a map which has a list of reservations
-                        Map<String, Object> reservationMap = doc.getData();
-                        List<HashMap<String, Object>> reservationMapList = (List<HashMap<String, Object>>) reservationMap.get("reservations");
-                        for (Map<String, Object> mapReservation : reservationMapList) {
-                            // Its a map in the database
-                            Map.Entry<String, Object> entry = mapReservation.entrySet().iterator().next();
-                            Reservation reservation = new Reservation();
-                            reservation.setReservationFromMap(entry);
-                            if (!reservation.getOfferID().equals(offer.getOfferID())) {
-                                newReservationList.add(reservation);
-                            }
-                        }
-                        saveReservationList(newReservationList, "reservations", doc.getId(), context);
-                    }
-                }
-            }
-        });
-    }
-
-    private Task<QuerySnapshot> getAllReservations() {
-        return FirebaseFirestore.getInstance().collection("reservations").get();
-    }
-
-    public void deleteReservationsForOffer(Offer offer, Context context, String collection, String document) {
-        List<Reservation> newListReservation = new ArrayList<>();
-        getListOfReservations(collection, document).addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                if (task.isSuccessful()) {
-                    List<HashMap<String, Object>> reservationList = (List<HashMap<String, Object>>) task.getResult().get("reservations");
-                    for (HashMap<String, Object> reservationMap : reservationList) {
-                        for (Map.Entry<String, Object> entry : reservationMap.entrySet()) {
-                            Reservation reservation = (Reservation) entry.getValue();
-                            if (!reservation.getOfferID().equals(offer.getOfferID())) {
-                                newListReservation.add(reservation);
-                            }
-                        }
-                    }
-                    saveReservationList(newListReservation, "reservations", FirebaseAuth.getInstance().getUid(), context);
-                } else {
-                    if (mDialog.isShowing()) {
-                        mDialog.dismiss();
-                    }
-                }
-            }
-        });
-    }
-
-    private Task<Void> saveNewReservationList(List<Reservation> newListReservation, String collection, String docID) {
-        return FirebaseFirestore.getInstance().collection(collection).document(docID).update("reservations", newListReservation);
-    }
-
-
-    public void deleteReservationManager(Reservation reservation, Context context, String collection, String document) {
-        displayLoadingDialog(context, R.string.frag_displayOffer_dialog_delete_reservation_message);
-        List<Reservation> newListReservation = new ArrayList<>();
-        getListOfReservations(collection, document).addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                if (task.isSuccessful()) {
-                    final List<Map<String, Object>> reservationList = (List<Map<String, Object>>) task.getResult().get("reservations");
-                    for (Map<String, Object> reservationMap : reservationList) {
-                        Map.Entry<String, Object> entry = reservationMap.entrySet().iterator().next();
-                        Reservation reservationCur = new Reservation();
-                        reservationCur.setReservationFromMap(entry);
-                        if (!reservationCur.getID().equals(reservation.getID())) {
-                            newListReservation.add(reservationCur);
-                        }
-
-                    }
-                    saveReservationList(newListReservation, collection, document, context);
-                    deleteReservationUser(reservation, context, "reservations", reservation.getClientID());
-                } else {
-                    if (mDialog.isShowing()) {
-                        mDialog.dismiss();
-                    }
-                }
-            }
-        });
-
-    }
-
-    private void deleteReservationUser(Reservation reservation, Context context, String collection, String document) {
-        List<Reservation> newListReservation = new ArrayList<>();
-        getListOfReservations(collection, document).addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                if (task.isSuccessful()) {
-                    List<HashMap<String, Object>> reservationList = (List<HashMap<String, Object>>) task.getResult().get("reservations");
-                    for (HashMap<String, Object> reservationMap : reservationList) {
-                        Map.Entry<String, Object> entry = reservationMap.entrySet().iterator().next();
-                        Reservation reservationCur = new Reservation();
-                        reservationCur.setReservationFromMap(entry);
-                        if (!reservationCur.getID().equals(reservation.getID())) {
-                            newListReservation.add(reservationCur);
-                        }
-
-                    }
-                    saveReservationList(newListReservation, collection, document, context);
-                } else {
-                    if (mDialog.isShowing()) {
-                        mDialog.dismiss();
-                    }
-                }
-            }
-        });
-    }
-
-    private void saveReservationList(List<Reservation> newListReservation, String collection, String document, Context context) {
-        saveNewReservationList(newListReservation, collection, document).addOnCompleteListener(new OnCompleteListener<Void>() {
-            @Override
-            public void onComplete(@NonNull Task<Void> task) {
-                if (!task.isSuccessful()) {
-                    Toast.makeText(context, "Reservation(s) deleted!", Toast.LENGTH_LONG).show();
-                }
-                if (mDialog.isShowing()) {
-                    mDialog.dismiss();
-                }
-            }
-        });
-    }
-
-    private Task<DocumentSnapshot> getListOfReservations(String collection, String document) {
-        return FirebaseFirestore.getInstance().collection(collection).document(document).get();
-    }
 
     private Task<Void> deleteOffer(Offer offer) {
         return FirebaseFirestore.getInstance().collection("offers").document(offer.getOfferID()).delete();
@@ -200,8 +68,8 @@ public class OfferActions {
             @Override
             public void onComplete(@NonNull Task<Void> task) {
                 if (task.isSuccessful()) {
-                    Reservation reservation = createReservation(offer, totalPrice);
-                    saveReservationToFirebase(reservation, context, offer.getManagerID());
+                    Reservation reservation = ReservationActions.getInstance().createReservation(offer, totalPrice);
+                    ReservationActions.getInstance().saveReservationToFirebase(reservation, context, offer.getManagerID(), mDialog);
                 } else {
                     mDialog.dismiss();
                     Toast.makeText(context, "Error while making reservation", Toast.LENGTH_LONG).show();
@@ -211,63 +79,12 @@ public class OfferActions {
 
     }
 
-    private Reservation createReservation(Offer offer, double totalPrice) {
-        Reservation reservation = new Reservation();
-        reservation.setEndDate(offer.getDateEnd());
-        reservation.setStartDate(offer.getDateStart());
-        reservation.setLocation(offer.getCityName());
-        reservation.setOfferID(offer.getOfferID());
-        reservation.setPresentationURL(offer.getPresentationURL());
-        reservation.setPrice(String.valueOf(totalPrice));
-        return reservation;
-    }
 
     private void displayLoadingDialog(Context context, int messageResource) {
         mDialog = DialogShow.getInstance().getDisplayDialog(context, messageResource);
         mDialog.show();
     }
 
-    private void saveReservationToFirebase(Reservation reservation, Context context, String managerID) {
-        Map<String, Object> currentReservation = new HashMap<>();
-        reservation.setClientID(FirebaseAuth.getInstance().getCurrentUser().getUid());
-        reservation.setID(String.valueOf(UUID.randomUUID()));
-        currentReservation.put(reservation.getID(), reservation);
-        FirebaseFirestore.getInstance()
-                .collection("reservations")
-                .document(reservation.getClientID())
-                .update("reservations", FieldValue.arrayUnion(currentReservation))
-                .addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        if (task.isSuccessful()) {
-                            addReservationToManager(reservation, managerID, context);
-
-                        } else {
-                            Toast.makeText(context, "Reservation Error!" + task.getException().toString(), Toast.LENGTH_LONG).show();
-                        }
-                    }
-                });
-    }
-
-    private void addReservationToManager(Reservation reservation, String managerID, Context context) {
-        Map<String, Object> currentReservation = new HashMap<>();
-        currentReservation.put(reservation.getID(), reservation);
-        FirebaseFirestore.getInstance()
-                .collection("reservationsManager")
-                .document(managerID)
-                .update("reservations", FieldValue.arrayUnion(currentReservation))
-                .addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        mDialog.dismiss();
-                        if (task.isSuccessful()) {
-                            Toast.makeText(context, "Reservation Made!", Toast.LENGTH_LONG).show();
-                        } else {
-                            Toast.makeText(context, "Reservation Error!" + task.getException().toString(), Toast.LENGTH_LONG).show();
-                        }
-                    }
-                });
-    }
 
     private Task<Void> updateOfferInFirebase(Offer offer) {
         return FirebaseFirestore.getInstance()
